Allocator

+++ src/heap/read-only-spaces.cc
@@ -612,7 +612,8 @@ void ReadOnlySpace::EnsureSpaceForAllocation(int size_in_bytes) {
 HeapObject ReadOnlySpace::TryAllocateLinearlyAligned(
     int size_in_bytes, AllocationAlignment alignment) {
   Address current_top = top_;
-  int filler_size = Heap::GetFillToAlign(current_top, alignment);
+  int filler_size =
+      static_cast<int>(Heap::GetFillToAlign(current_top, alignment));
 
   Address new_top = current_top + filler_size + size_in_bytes;
   if (new_top > limit_) return HeapObject();
